re-factor: modularized overall committed
- The Driver code is in `todo.py`, but the TodoList Class definition is separated in another file `TodoList.py`
- Changed how we deal with `try..except` in `todo.py` by wrapping it around the entire `if..else` statements

issue #19
