+++ TodoList.py
@@ -1,0 +1,38 @@
+class TodoList:
+    def __init__(self):
+        self.tasks = []
+
+    def add_task(self, task):
+        """Adds a new task to the list."""
+        self.tasks.append(task)
+        print(f'Task added: {task}')
+
+    def list_tasks(self):
+        """Lists all tasks in the to-do list."""
+        if not self.tasks:
+            print("No tasks in the list!")
+        else:
+            for idx, task in enumerate(self.tasks, start=1):
+                print(f'{idx}. {task}')
+
+    def add_task_date(self, task_number, due_date):
+        """Add a due date to a task."""
+        if task_number <= 0 or task_number > len(self.tasks):
+            print("Invalid task number!")
+        else:
+            task = self.tasks[task_number - 1]
+            if type(task) is str:
+                self.tasks[task_number - 1] = (task, due_date)
+            else:
+                task_name = task[0]
+                self.tasks[task_number - 1] = (task_name, due_date)
+            
+            print(f'Task updated: {self.tasks[task_number - 1]}')
+
+    def delete_task(self, task_number):
+        """Deletes a task by its number in the list."""
+        if task_number <= 0 or task_number > len(self.tasks):
+            print("Invalid task number!")
+        else:
+            removed_task = self.tasks.pop(task_number - 1)
+            print(f'Task removed: {removed_task}')
